Extract composer order visuals
