+++ scripts/test-system-release-package.sh
@@ -128,6 +128,11 @@ if ! grep -qx "press-system-${version}/assets/js/composer-order-diff-ui.js" "${e
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/composer-order-visual.js" "${entries_file}"; then
+  echo "expected package to include composer order visual code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/composer-index-tabs-ui.js" "${entries_file}"; then
   echo "expected package to include composer index/tabs UI code" >&2
   exit 1
